add env & split some file 🚀
