+++ package.json
@@ -8,7 +8,9 @@
     "dev": "nodemon index.js",
     "start": "node index.js",
     "build:local": "docker build -t node-socket .",
-    "start:docker": "docker run -d --network tunnel -p 3000:3000 --name node-socket node-socket"
+    "start:docker": "docker run -d --network tunnel -p 3000:3000 --name node-socket node-socket",
+    "format:check": "prettier . --check --config .prettierrc",
+    "format:all": "prettier . --write"
   },
   "keywords": [],
   "author": "",
@@ -18,6 +20,7 @@
   },
   "dependencies": {
     "@fastify/static": "^7.0.1",
+    "dotenv": "^16.4.5",
     "fastify": "^4.26.1",
     "fastify-socket.io": "^5.0.0",
     "socket.io": "^4.7.4"
