+++ routes/user.js
@@ -1,0 +1,5 @@
+export default async function (fastify, opts) {
+  fastify.get('/user', async (req, reply) => {
+    return '/user';
+  });
+}
